+++ server/src/constants.ts
@@ -3,6 +3,7 @@ export enum CustomMessages {
 	ProjectIndexed = 'choicescript/projectindexed',
 	SelectionWordCountRequest = 'choicescript/selectionwordcount',
 	UpdatedSceneFilesPath = 'choicescript/scenefilespath',
+	UpdatedMyGamePath = 'choicescript/mygamepath',
 	UpdatedWordCount = 'choicescript/updatedwordcount',
 	WordCountRequest = 'choicescript/wordcount',
 }